+++ src/sargasso/common/counter.h
@@ -1,7 +1,7 @@
 #ifndef SARGASSO_COMMON_COUNTER_H
 #define SARGASSO_COMMON_COUNTER_H
 
-#include <cstdint>
+#include <cstdlib>
 
 namespace sargasso {
 namespace common {
@@ -11,15 +11,15 @@ class Counter {
     void reset();
     void increase();
     void decrease();
-    bool operator==(uint64_t n) const;
-    bool operator>(uint64_t n) const;
-    bool operator>=(uint64_t n) const;
-    bool operator<=(uint64_t n) const;
-    bool operator<(uint64_t n) const;
-    uint64_t get() const;
+    bool operator==(size_t n) const;
+    bool operator>(size_t n) const;
+    bool operator>=(size_t n) const;
+    bool operator<=(size_t n) const;
+    bool operator<(size_t n) const;
+    size_t get() const;
 
    private:
-    uint64_t _count = 0;
+    size_t _count = 0;
 };
 
 }  // namespace common
